Remove GrantListWithFilter and embed logic directly into GrantList
